220723 3주차 MR

See merge request sw_track/class_02_seoul/web_project_2/team11/project-goodbye!76

+++ server/src/routes/remembrance-router.ts
@@ -1,42 +1,33 @@
 import { Router } from 'express';
 import {
-    updateCommentJoiSchema,
     createCommentJoiSchema,
-} from '../db/schemas/joi-schemas/comment-joi-schema';
-import { updateRemembranceJoiSchema } from '../db/schemas/joi-schemas/remembrance-joi-schema';
-import { loginRequired } from '../middlewares';
+    deleteCommentJoiSchema,
+    updateCommentJoiSchema,
+} from '../db/schemas/joi-schemas';
+import { ResComment, ResRemembrance } from '../db';
 import { remembranceService, commentService } from '../services';
 
 const remembranceRouter = Router();
 
-// 추모 데이터 생성 - 언제?
-remembranceRouter.post('/', loginRequired, async (req, res, next) => {
+// 전체 사망한 유저의 추모 데이터 조회
+remembranceRouter.get('/', async (req, res, next) => {
     try {
-        // loginRequired 이용해서 로그인한 사용자의 id 받아오기
-        if (!req.user) {
-            throw new Error('로그인 먼저 진행해 주세요.');
-        }
-        const { _id: userId } = req.user;
-        const { dateOfDeath } = req.body; // death 받아오나? - 생성 시점에 따라 다를듯 일단 받아오기
+        const remembrances: ResRemembrance[] =
+            await remembranceService.getRemembrances();
 
-        const newRemembrance = await remembranceService.addRemembrance(
-            userId,
-            dateOfDeath,
-        );
-
-        res.status(201).json(newRemembrance);
+        res.status(200).json(remembrances);
     } catch (error) {
         next(error);
     }
 });
 
 // 최근 업데이트된 추모 조회
-remembranceRouter.get('/', async (req, res, next) => {
+remembranceRouter.get('/recent', async (req, res, next) => {
     try {
         // query로 받거나 8
-        const count = Number(req.query.count) || 8;
+        const count: number = Number(req.query.count) || 8;
 
-        const recentRemembrances =
+        const recentRemembrances: ResRemembrance[] =
             await remembranceService.getRecentRemembrances(count);
 
         res.status(200).json(recentRemembrances);
@@ -45,14 +36,13 @@ remembranceRouter.get('/', async (req, res, next) => {
     }
 });
 
-// 특정 추모 조회
+// remembranceId로 특정 추모 조회
 remembranceRouter.get('/:remembranceId', async (req, res, next) => {
     try {
         const { remembranceId } = req.params;
 
-        const remembrance = await remembranceService.getRemembranceById(
-            remembranceId,
-        );
+        const remembrance: ResRemembrance =
+            await remembranceService.getRemembranceById(remembranceId);
 
         res.status(200).json(remembrance);
     } catch (error) {
@@ -60,14 +50,16 @@ remembranceRouter.get('/:remembranceId', async (req, res, next) => {
     }
 });
 
-// 특정 추모글 조회
+// 특정 추모글 조회 -> remembranceId를 사용하지 않는데 따로 빼는게 좋을까?
 remembranceRouter.get(
     '/:remembranceId/comments/:commentId',
     async (req, res, next) => {
         try {
             const { commentId } = req.params;
 
-            const comment = await commentService.getCommentById(commentId);
+            const comment: ResComment = await commentService.getCommentById(
+                commentId,
+            );
 
             res.status(200).json(comment);
         } catch (error) {
@@ -76,55 +68,23 @@ remembranceRouter.get(
     },
 );
 
-// 추모 수정 - fullName, dateOfBirth, photo는 user data -> 언제 어떻게 변경?
-remembranceRouter.patch('/:remembranceId', async (req, res, next) => {
-    try {
-        const { remembranceId } = req.params;
-        const { fullName, dateOfBirth, dateOfDeath, isPublic, photo } =
-            req.body;
-        const isValid = await updateRemembranceJoiSchema.validateAsync({
-            fullName,
-            dateOfBirth,
-            dateOfDeath,
-            isPublic,
-            photo,
-        });
-
-        const remembrance = await remembranceService.setRemembrance(
-            remembranceId,
-            {
-                ...(fullName && { fullName }),
-                ...(dateOfBirth && { dateOfBirth }),
-                ...(dateOfDeath && { dateOfDeath }),
-                ...(typeof isPublic === 'boolean' && { isPublic }),
-                ...(photo && { photo }),
-            },
-        );
-
-        res.status(201).json(remembrance);
-    } catch (error) {
-        next(error);
-    }
-});
-
 // 추모 글 추가
 remembranceRouter.post('/:remembranceId/comments', async (req, res, next) => {
     try {
         const { remembranceId } = req.params;
         const { writer, title, content, password } = req.body;
-        const isValid = await createCommentJoiSchema.validateAsync({
-            writer,
-            title,
-            content,
-            password,
-        });
 
-        const newComment = await commentService.addComment(remembranceId, {
+        const commentInfo = {
             writer,
             title,
             content,
             password,
-        });
+        };
+        await createCommentJoiSchema.validateAsync(commentInfo);
+        const newComment: ResComment = await commentService.addComment(
+            remembranceId,
+            commentInfo,
+        );
 
         res.status(201).json(newComment);
     } catch (error) {
@@ -139,21 +99,17 @@ remembranceRouter.patch(
         try {
             const { commentId } = req.params;
             const { writer, title, content, password } = req.body;
-            const isValid = await updateCommentJoiSchema.validateAsync({
-                writer,
-                title,
-                content,
-                password,
-            });
 
-            const comment = await commentService.setCommet(
-                commentId,
+            const commentInfo = {
+                ...(writer && { writer }),
+                ...(title && { title }),
+                ...(content && { content }),
                 password,
-                {
-                    ...(writer && { writer }),
-                    ...(title && { title }),
-                    ...(content && { content }),
-                },
+            };
+            await updateCommentJoiSchema.validateAsync(commentInfo);
+            const comment: ResComment = await commentService.setCommet(
+                commentId,
+                commentInfo,
             );
 
             res.status(200).json(comment);
@@ -169,13 +125,18 @@ remembranceRouter.delete(
     async (req, res, next) => {
         try {
             const { remembranceId, commentId } = req.params;
-            const { password } = req.body;
+            const password = req.get('password');
+            if (!password) {
+                throw new Error('비밀번호가 필요합니다.');
+            }
 
-            const result = await commentService.deleteComment(
-                remembranceId,
-                commentId,
-                password,
-            );
+            await deleteCommentJoiSchema.validateAsync(password);
+            const result: { result: string } =
+                await commentService.deleteComment(
+                    remembranceId,
+                    commentId,
+                    password,
+                );
 
             res.status(201).json(result);
         } catch (error) {
